tests/integration/requirements.txt
blockfrost-python
connexion==2.9.0
flask-cors==3.0.10
Flask-Migrate==3.1.0
flask_sqlalchemy==2.5.1
jsonschema
markupsafe==2.0.1
psycopg2-binary==2.9.3
pycardano
pytest==7.1.1
pytest-cov
pytest-mock==3.6.1
python-dotenv==0.19.2
sqlalchemy==1.4.27
swagger-ui-bundle==0.0.9
